new updates for security and bug fixes

+++ php/deletePost.php
@@ -2,7 +2,13 @@
     require("../database.php");
     $database = new Database();
     $err_message = 1;
-    $post_id = $_POST['post_id'];
+    $post_id = NULL;
+    if (isset($_POST['post_id'])) {
+        $post_id = $_POST['post_id'];
+    }
+    if (!isset($post_id)) {
+        return;
+    }
     $database->where("post_id", $post_id);
     $deletePost = $database->deleteRow("posts");
     if (!$deletePost) {
